auth_service/requirements.txt
# ---- Framework ----
fastapi
uvicorn[standard]

# ---- Base de Datos (MariaDB) ----
sqlalchemy
pymysql

# ---- Seguridad (Contraseñas y Tokens) ----
passlib[bcrypt]
python-jose[cryptography]

# ---- Validación de Datos ----
pydantic
email-validator

# ---- Cliente HTTP (para llamar a otros servicios) ----
